# docs/openapi/novel-platform.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Novel Platform API",
    "version": "0.1.0",
    "description": "Versioned REST contract for the novel platform backend. Browser clients use the same-origin Next BFF: replace the backend prefix `/api/v1` with `/api/novel` for the `public`, `account`, `author`, and `admin` route groups. The BFF owns browser cookies, CSRF validation, and the upstream internal headers; do not call protected backend routes from browser JavaScript directly. `/api/v1/auth/*` is an internal BFF-to-backend contract and is intentionally not exposed by the catch-all browser proxy. This contract describes implemented local platform behavior only. Phone/email verification, WeChat/QQ OAuth, external payment, and production Qwen credentials are not represented as available API integrations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Backend service. Protected and auth routes require BFF-only headers."
    }
  ],
...
  "paths": {
...
    "/api/v1/author/books": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorBooks",
        "summary": "List books owned by the current author",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/BookList",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Author workspace"],
        "operationId": "createAuthorBook",
        "summary": "Create an author-owned draft book",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Book",
        "requestBody": { "$ref": "#/components/requestBodies/BookRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/author/books/{bookId}": {
      "put": {
        "tags": ["Author workspace"],
        "operationId": "updateAuthorBook",
        "summary": "Update author-owned book metadata",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Book",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/BookUpdateRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["Author workspace"],
        "operationId": "deleteAuthorBook",
        "summary": "Delete an author-owned book",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/DeleteResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/volumes": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorVolumes",
        "summary": "List volumes for an author-owned book",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/VolumeList",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "post": {
        "tags": ["Author workspace"],
        "operationId": "createAuthorVolume",
        "summary": "Create a volume for an author-owned book",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Volume",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/VolumeRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/chapters": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorChapters",
        "summary": "List chapters for an author-owned book",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/ChapterList",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "post": {
        "tags": ["Author workspace"],
        "operationId": "createAuthorChapter",
        "summary": "Create a chapter, optionally submitting it to automatic screening",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Chapter",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/ChapterRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/volumes/{volumeId}/chapters": {
      "post": {
        "tags": ["Author workspace"],
        "operationId": "createVolumeDraftChapter",
        "summary": "Create a draft chapter in an author-owned volume",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Chapter",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/PathVolumeId" }],
        "requestBody": { "$ref": "#/components/requestBodies/DraftChapterRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/chapters/{chapterId}": {
      "put": {
        "tags": ["Author workspace"],
        "operationId": "updateAuthorChapter",
        "summary": "Update an author-owned chapter",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Chapter",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/PathChapterId" }],
        "requestBody": { "$ref": "#/components/requestBodies/ChapterUpdateRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["Author workspace"],
        "operationId": "deleteAuthorChapter",
        "summary": "Delete an author-owned chapter",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/DeleteResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/PathChapterId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/chapters/{chapterId}/submit": {
      "post": {
        "tags": ["Author workspace"],
        "operationId": "submitAuthorChapter",
        "summary": "Submit an author-owned chapter to automatic screening",
        "description": "The returned normalized status can show a review-required state. This is not a browser-accessible Qwen provider API.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Chapter",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/PathChapterId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
    "/api/v1/author/books/{bookId}/chapters/{chapterId}/schedule": {
      "post": {
        "tags": ["Author workspace"],
        "operationId": "scheduleAuthorChapter",
        "summary": "Schedule a future chapter publication attempt",
        "description": "The due attempt still follows the configured automatic screening path; scheduling does not bypass moderation.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Chapter",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/PathChapterId" }],
        "requestBody": { "$ref": "#/components/requestBodies/ScheduleChapterRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/author/scheduled-publications/run": {
      "post": {
        "tags": ["Author workspace"],
        "operationId": "runDueAuthorPublications",
        "summary": "Run due publication processing for the current author",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/DuePublicationResult",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/author/books/{bookId}/submit": {
      "post": {
        "tags": ["Author workspace"],
        "operationId": "submitAuthorBook",
        "summary": "Submit a complete author-owned book for administrator review",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/Book",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/author/books/{bookId}/comments": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorBookComments",
        "summary": "Page all comment states for an author-owned book",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/CommentPageResponse",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/InteractionStatus" }, { "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/InteractionPageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/author/books/{bookId}/annotations": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorBookAnnotations",
        "summary": "Page all paragraph annotation states for an author-owned book",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/AnnotationPageResponse",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/InteractionStatus" }, { "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/InteractionPageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "novel_session",
        "description": "Opaque browser BFF session cookie. The browser never receives the backend session id."
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-CSRF",
        "description": "Required by the BFF on unsafe browser requests and must match its CSRF cookie/session binding."
      },
      "InternalBffKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Internal-Key",
        "description": "Private BFF-to-backend credential. Never send this header from a browser."
      },
      "BackendBffSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Bff-Session",
        "description": "Server-side backend session credential held by the BFF only. Never expose it to browser code."
      }
    },
    "parameters": {
...
      "PathBookId": { "name": "bookId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "PathChapterId": { "name": "chapterId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "PathVolumeId": { "name": "volumeId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0, "maximum": 100000 } },
      "InteractionPageSize": { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1 }, "description": "The controller accepts an interaction page size; callers should use a bounded value." },
...
      "InteractionStatus": { "name": "status", "in": "query", "schema": { "type": "string" }, "description": "Optional stored comment or annotation status filter." },
...
    },
    "requestBodies": {
...
      "BookRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BookRequest" } } } },
      "BookUpdateRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BookUpdateRequest" } } } },
      "VolumeRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VolumeRequest" } } } },
      "ChapterRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChapterRequest" } } } },
      "DraftChapterRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DraftChapterRequest" } } } },
      "ChapterUpdateRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChapterUpdateRequest" } } } },
      "ScheduleChapterRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleChapterRequest" } } } },
...
    },
    "responses": {
      "Success": { "description": "Successful platform envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } } },
      "BadRequest": { "description": "Invalid request envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "Missing, expired, or invalid BFF session/internal BFF credential.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "Authenticated principal lacks the required role or ownership.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "Requested resource does not exist or is not visible to the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Requested state transition conflicts with current durable state.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unavailable": { "description": "A required internal dependency is unavailable.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer", "example": 200 }, "msg": { "type": "string", "example": "ok" }, "data": {} }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer" }, "msg": { "type": "string" }, "data": { "type": "null" } }
      },
      "Book": {
        "type": "object",
        "required": ["id", "title", "author", "category", "words", "serialStatus", "synopsis", "cover", "status", "authorId", "heat", "purchasePrice"],
        "properties": {
          "id": { "type": "integer", "format": "int64" }, "title": { "type": "string" }, "author": { "type": "string" }, "category": { "type": "string" }, "words": { "type": "integer" }, "serialStatus": { "type": "string" }, "synopsis": { "type": "string" }, "cover": { "type": "string" }, "status": { "type": "string", "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED", "NEEDS_REVIEW", "REJECTED", "OFFLINE"] }, "authorId": { "type": "integer", "format": "int64" }, "heat": { "type": "integer", "format": "int64" }, "purchasePrice": { "type": "integer", "format": "int64", "description": "Server-owned platform-token amount." }
        }
      },
      "BookList": { "type": "array", "items": { "$ref": "#/components/schemas/Book" } },
      "Volume": { "type": "object", "required": ["id", "bookId", "title", "orderNo", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "title": { "type": "string" }, "orderNo": { "type": "integer" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "VolumeList": { "type": "array", "items": { "$ref": "#/components/schemas/Volume" } },
      "Chapter": { "type": "object", "required": ["id", "bookId", "title", "content", "published", "status", "orderNo"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "volumeId": { "type": ["integer", "null"], "format": "int64" }, "title": { "type": "string" }, "content": { "type": "string" }, "published": { "type": "boolean" }, "status": { "type": "string", "enum": ["DRAFT", "SCHEDULED", "PUBLISHED", "NEEDS_REVIEW"] }, "scheduledPublishAt": { "type": ["string", "null"], "format": "date-time" }, "publishedAt": { "type": ["string", "null"], "format": "date-time" }, "reviewReason": { "type": "string" }, "orderNo": { "type": "integer" } } },
      "ChapterList": { "type": "array", "items": { "$ref": "#/components/schemas/Chapter" } },
      "Comment": { "type": "object", "required": ["id", "bookId", "userId", "authorName", "content", "status", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": ["integer", "null"], "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "content": { "type": "string" }, "status": { "type": "string", "enum": ["PENDING_REVIEW", "VISIBLE", "REJECTED"] }, "createdAt": { "type": "string", "format": "date-time" } } },
      "CommentPageResponse": { "type": "object", "required": ["items", "meta"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } }, "meta": { "$ref": "#/components/schemas/PageMeta" } } },
      "ParagraphAnnotation": { "type": "object", "required": ["id", "bookId", "chapterId", "userId", "authorName", "paragraphIndex", "selectionStart", "selectionEnd", "selectedText", "note", "shareIntent", "status", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": "integer", "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "paragraphIndex": { "type": "integer" }, "selectionStart": { "type": "integer" }, "selectionEnd": { "type": "integer" }, "selectedText": { "type": "string" }, "note": { "type": "string" }, "shareIntent": { "type": "boolean" }, "status": { "type": "string", "enum": ["PRIVATE", "PENDING_REVIEW", "VISIBLE", "REJECTED"] }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "AnnotationPageResponse": { "type": "object", "required": ["items", "meta"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/ParagraphAnnotation" } }, "meta": { "$ref": "#/components/schemas/PageMeta" } } },
...
      "PageMeta": { "type": "object", "required": ["total", "page", "size"], "properties": { "total": { "type": "integer", "format": "int64", "minimum": 0 }, "page": { "type": "integer", "minimum": 0 }, "size": { "type": "integer", "minimum": 1 } } },
...
      "DeleteResult": { "type": "object", "required": ["id", "deleted"], "properties": { "id": { "type": "integer", "format": "int64" }, "deleted": { "type": "boolean", "const": true } } },
      "DuePublicationResult": { "type": "object", "required": ["processed", "published", "needsReview"], "properties": { "processed": { "type": "integer", "minimum": 0 }, "published": { "$ref": "#/components/schemas/ChapterList" }, "needsReview": { "$ref": "#/components/schemas/ChapterList" } } },
...
      "BookRequest": { "type": "object", "required": ["title", "category", "synopsis"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 }, "category": { "type": "string", "minLength": 1, "maxLength": 128 }, "synopsis": { "type": "string", "minLength": 1, "maxLength": 20000 } }, "additionalProperties": false },
      "BookUpdateRequest": { "type": "object", "required": ["title", "category", "synopsis"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 }, "category": { "type": "string", "minLength": 1, "maxLength": 128 }, "synopsis": { "type": "string", "minLength": 1, "maxLength": 20000 }, "serialStatus": { "type": "string", "maxLength": 32 }, "cover": { "type": "string", "maxLength": 1024 } }, "additionalProperties": false },
      "VolumeRequest": { "type": "object", "required": ["title"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 } }, "additionalProperties": false },
      "ChapterRequest": { "type": "object", "required": ["title", "content", "submit"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 }, "content": { "type": "string", "minLength": 1, "maxLength": 20000 }, "submit": { "type": "boolean" }, "volumeId": { "type": ["integer", "null"], "format": "int64" } }, "additionalProperties": false },
      "DraftChapterRequest": { "type": "object", "required": ["title", "content"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 }, "content": { "type": "string", "minLength": 1, "maxLength": 20000 } }, "additionalProperties": false },
      "ChapterUpdateRequest": { "type": "object", "required": ["title", "content"], "properties": { "title": { "type": "string", "minLength": 1, "maxLength": 255 }, "content": { "type": "string", "minLength": 1, "maxLength": 20000 }, "volumeId": { "type": ["integer", "null"], "format": "int64" } }, "additionalProperties": false },
      "ScheduleChapterRequest": { "type": "object", "required": ["publishAt"], "properties": { "publishAt": { "type": "string", "format": "date-time", "description": "A future instant." } }, "additionalProperties": false },
...
    }
  }
}